Extract IpodModernPlayPauseIcon to shared screen module
