+++ src/apps/ipod/components/screen/index.ts
@@ -4,3 +4,4 @@ export { Scrollbar } from "./Scrollbar";
 export { MenuListItem } from "./MenuListItem";
 export { ScrollingText } from "./ScrollingText";
 export { StatusDisplay } from "./StatusDisplay";
+export { IpodModernPlayPauseIcon } from "./IpodModernPlayPauseIcon";
